requirements.txt
Flask==3.0.0
pandas==2.1.1
requests==2.31.0
scikit-learn==1.3.1
python-dotenv==1.0.0